chore: code style updates for cmd/image-builder
