+++ cmd/image-builder/cmd_haproxy.go
