+++ cmd/image-builder/cmd_kubeadm.go
@@ -9,6 +9,11 @@ import (
 )
 
 var (
+	kubeadmCfg struct {
+		kubernetesVersion string
+		pullExtraImages   []string
+	}
+
 	kubeadmCmd = &cobra.Command{
 		Use:          "kubeadm",
 		GroupID:      "build",
